+++ modules/Discipline/StudentFieldBreakdown.php
@@ -391,8 +391,6 @@ if(empty($_REQUEST['modfunc']))
 			}	
 ?>
 			<div id="chart"></div>
-			<script src="assets/js/colorbox/jquery.colorbox-min.js"></script>
-			<link rel="stylesheet" href="assets/js/colorbox/colorbox.css" type="text/css" media="screen" />
 			<script src="assets/js/jquery.jqplottocolorbox.js"></script>
 <?php
 			unset($_REQUEST['_ROSARIO_PDF']);