+++ modules/School_Setup/includes/Modules.inc.php
@@ -257,8 +257,6 @@ function _makeReadMe($module_title,$activated=null)
 		// convert MarkDown text to HTML
 		$readme_content = '<div class="markdown-to-html">' . $readme_content . '</div>';
 
-		$return .= includeOnceColorBox();
-		
 		$return .= '<div style="display:none;"><div id="README_'.$module_title.'">'.$readme_content.'</div></div>';
 
 		$return .= '<a class="colorboxinline" href="#README_'.$module_title.'">'.$module_title_echo.'</a>';