+++ modules/misc/Portal.php
@@ -545,9 +545,9 @@ function _formatContent($value,$column)
 	//FJ responsive rt td too large
 	//FJ Portal Assignments
 	if ( isset( $THIS_RET['COURSE'] ) )
-		$return = includeOnceColorBox('divAssignmentContent'.$id).'<DIV id="divAssignmentContent'.$id.'" class="rt2colorBox">';
+		$return = '<DIV id="divAssignmentContent'.$id.'" class="rt2colorBox">';
 	else
-		$return = includeOnceColorBox('divNoteContent'.$id).'<DIV id="divNoteContent'.$id.'" class="rt2colorBox">';
+		$return = '<DIV id="divNoteContent'.$id.'" class="rt2colorBox">';
 
 	// convert MarkDown to HTML
 	$return .= '<div class="markdown-to-html">' . $value . '</div>';